+++ arbitragelab/network/imports.py
@@ -34,7 +34,7 @@ class Golem:
         import arbitragelab.tearsheet as tearsheet
         import arbitragelab.time_series_approach as time_series_approach
         import arbitragelab.hedge_ratios as hedge_ratios
-        import arbitragelab.pairs_selection as pairs_selection
+        import arbitragelab.spread_selection as pairs_selection
         import arbitragelab.util as util
         segment.track('Import')