+++ docs/source/changelog.rst
@@ -12,6 +12,16 @@ Changelog
 ..
     For Help: https://releases.readthedocs.io/en/latest/index.html
 
+* :support:`71` Added Raw Docs for debugging to documentation.
+* :support:`71` Added UML Diagram for debugging to documentation.
+* :feature:`70` Updated requirements versions to the newest stable numpy, pandas etc.
+* :bug:`70 major` Fixed package breaking due to faulty werkzeug version.
+* :bug:`69 major` Fixed SCS package version breaking Sparse MR Module.
+* :feature:`68` Updated Cointegration Pairs Selection Module to work with any type of spread (3-leg, N-leg spread).
+* :feature:`68` Updated Hedge Ratios Module to work with any type of spread (3-leg, N-leg spread).
+* :feature:`68` Added Johansen Eigenvector, Box-Tiao Canonical Decomposition, Minimum ADF Test T-statistic methods to Hedge Ratios Module.
+* :support:`68` Added Spread Selection Tools Module to the documentation.
+* :support:`68` Reflected changes to Hedge Ratios Module in the documentation.
 * :support:`67` Added blog post links to documentation.
 * :support:`66` Added presentation slides and videos to documentation.